Refactor Error Throwing to Results

+++ Sources/HatTip/Client+URLSession.swift
@@ -2,7 +2,7 @@ import Foundation
 
 extension URLSession: Client {
 
-    public func send(_ request: Request, completion: @escaping (Result<Response, Error>) -> Void) {
+    public func send(_ request: Request, completion: @escaping (Result<Response, BasicError>) -> Void) {
         switch request.body {
         case .none, .some(.data):
             switch request.responseBodyHint {
@@ -18,11 +18,17 @@ extension URLSession: Client {
 
     private func sendViaDataTask(
         _ request: Request,
-        completion: @escaping (Result<Response, Error>) -> Void
+        completion: @escaping (Result<Response, BasicError>) -> Void
         ) {
 
         let task = self.dataTask(with: request.urlRequest) { (data, response, error) in
-            completion(.init { try .init(data: data, response: response, error: error) })
+            completion(
+                Response.make(
+                    data: data,
+                    response: response,
+                    error: error as NSError?
+                )
+            )
         }
         task.resume()
     }
@@ -30,23 +36,35 @@ extension URLSession: Client {
     private func sendViaDownloadTask(
         _ request: Request,
         to destination: URL?,
-        completion: @escaping (Result<Response, Error>) -> Void
+        completion: @escaping (Result<Response, BasicError>) -> Void
         ) {
 
         let task = self.downloadTask(with: request.urlRequest) { (tmpUrl, response, error) in
-            completion(
-                .init {
-                    let url = try tmpUrl.map { (src: URL) -> URL in
-                        let dst = destination ?? FileManager.default
-                            .temporaryDirectory
-                            .appendingPathComponent(UUID().uuidString, isDirectory: false)
-                        try? FileManager.default.removeItem(at: dst)
-                        try FileManager.default.moveItem(at: src, to: dst)
-                        return dst
-                    }
-                    return try .init(file: url, response: response, error: error)
-                }
-            )
+            guard let tmpUrl = tmpUrl else {
+                return completion(
+                    Response.make(
+                        url: nil,
+                        response: response,
+                        error: error as NSError?
+                    )
+                )
+            }
+            do {
+                let dst = destination ?? FileManager.default
+                    .temporaryDirectory
+                    .appendingPathComponent(UUID().uuidString, isDirectory: false)
+                try? FileManager.default.removeItem(at: dst)
+                try FileManager.default.moveItem(at: tmpUrl, to: dst)
+                completion(
+                    Response.make(
+                        url: dst,
+                        response: response,
+                        error: error as NSError?
+                    )
+                )
+            } catch {
+                completion(.failure(.init(fileError: error as NSError)))
+            }
         }
         task.resume()
     }
@@ -54,11 +72,17 @@ extension URLSession: Client {
     private func sendViaUploadTask(
         _ request: Request,
         from source: URL,
-        completion: @escaping (Result<Response, Error>) -> Void
+        completion: @escaping (Result<Response, BasicError>) -> Void
         ) {
 
         let task = self.uploadTask(with: request.urlRequest, fromFile: source) { (data, response, error) in
-            completion(.init { try .init(data: data, response: response, error: error) })
+            completion(
+                Response.make(
+                    data: data,
+                    response: response,
+                    error: error as NSError?
+                )
+            )
         }
         task.resume()
     }
